add time interval function to get events
