+++ app/backend/routes/events.js
@@ -3,11 +3,6 @@ const router = express.Router();
 const eventController = require('../controllers/events')
 const {modelBinder} = require('../controllers/db')
 const { Event } = require('../models/event')
-/*
-  Get endpoint for events page.
-  Check controller function for more detail
-*/
-router.get('/fromAPI', modelBinder(Event, 'Event'), eventController.getEventsFromAPI)
 
 /*
   Get endpoint for events page.
